fix: address rate-limit review findings
