+++ src/polymarket/clients/secure.py
@@ -365,7 +365,7 @@ class SecureClient:
         credentials: ApiKeyCreds,
         api_key: ApiKey | None,
         logger: logging.Logger | None,
-        on_rate_limit_update: RateLimitUpdateListener | None = None,
+        on_rate_limit_update: RateLimitUpdateListener | None,
     ) -> Self:
         try:
             wallet_checksum = to_checksum_address(wallet)
