+++ src/polymarket/rate_limit.py
@@ -2,12 +2,9 @@
 
 from __future__ import annotations
 
-import math
 from collections.abc import Callable
 from dataclasses import dataclass
 from typing import TypeAlias
-
-import httpx
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -36,40 +33,6 @@ class RateLimitUpdate:
 
 RateLimitUpdateListener: TypeAlias = Callable[[RateLimitUpdate], None]
 """Listener invoked whenever a response reports per-signer rate-limit state."""
-
-
-def parse_rate_limit_headers(headers: httpx.Headers) -> RateLimitUpdate | None:
-    """Parse the ``Poly-RateLimit-*`` response headers.
-
-    Returns ``None`` when the response carries none of them.
-    """
-    remaining = _parse_numeric_header(headers.get("Poly-RateLimit-Remaining"))
-    reset = _parse_numeric_header(headers.get("Poly-RateLimit-Reset"))
-    tier = _parse_text_header(headers.get("Poly-RateLimit-Tier"))
-    warning_header = _parse_text_header(headers.get("Poly-RateLimit-Warning"))
-    warning = warning_header is not None and warning_header.lower() == "true"
-
-    if remaining is None and reset is None and tier is None and not warning:
-        return None
-
-    return RateLimitUpdate(remaining=remaining, reset=reset, tier=tier, warning=warning)
-
-
-def _parse_numeric_header(value: str | None) -> float | None:
-    if value is None:
-        return None
-    try:
-        parsed = float(value.strip())
-    except ValueError:
-        return None
-    return parsed if math.isfinite(parsed) else None
-
-
-def _parse_text_header(value: str | None) -> str | None:
-    if value is None:
-        return None
-    stripped = value.strip()
-    return stripped if stripped else None
 
 
 __all__ = ["RateLimitUpdate", "RateLimitUpdateListener"]
